requirements.txt
openai~=1.59.7
llama-index-core~=0.12.11
fastapi~=0.115.6
uvicorn~=0.34.0